Fetching profile data using useQuery and persisting in sessionstorage
The profile data is fetched from a local /api/profile which searches all the user domain data(student, instructor or organisaction)
based on the users domain. The data is then fetched using tanstack query for caching and persisting the cache in the session storage.
To get the data useUserProfile hook, the data consists of {...userData, instructor:{}, student:{}, organisation:{}, invalidateQuery}

invalidateQuery is a function that forces a refresh and recache of the data after an update